Move fs from its package to database package
